+++ cmd/root.go
@@ -9,8 +9,10 @@ import (
 )
 
 var (
-	cfgFile string
-	apiKey  string
+	cfgFile         string
+	apiKey          string
+	endpoint        string
+	defaultEndpoint = "https://api.honeybadger.io"
 )
 
 // rootCmd represents the base command when called without any subcommands
@@ -31,8 +33,13 @@ func init() {
 
 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is config/honeybadger.yml)")
 	rootCmd.PersistentFlags().StringVar(&apiKey, "api-key", "", "Honeybadger API key")
+	rootCmd.PersistentFlags().StringVar(&endpoint, "endpoint", defaultEndpoint, "Honeybadger endpoint")
+
 	if err := viper.BindPFlag("api_key", rootCmd.PersistentFlags().Lookup("api-key")); err != nil {
 		fmt.Printf("error binding api-key flag: %v\n", err)
+	}
+	if err := viper.BindPFlag("endpoint", rootCmd.PersistentFlags().Lookup("endpoint")); err != nil {
+		fmt.Printf("error binding endpoint flag: %v\n", err)
 	}
 }
 
@@ -48,6 +55,7 @@ func initConfig() {
 
 	viper.AutomaticEnv()
 	viper.SetEnvPrefix("HONEYBADGER")
+	viper.SetDefault("endpoint", defaultEndpoint)
 
 	if err := viper.ReadInConfig(); err == nil {
 		fmt.Fprintln(os.Stderr, "Using config file:", viper.ConfigFileUsed())